DOC: build API pages only on change

+++ docs/conf.py
@@ -1,4 +1,5 @@
 import datetime
+import filecmp
 import os
 import shutil
 
@@ -92,7 +93,6 @@ linkcheck_ignore = [
 
 
 # Copy API (sub-)module RST files to docs/api/ folder ---------------------
-audeer.rmdir('api')
 audeer.mkdir('api')
 api_src_files = audeer.list_file_names('api-src')
 api_dst_files = [
@@ -100,4 +100,8 @@ api_dst_files = [
     for src_file in api_src_files
 ]
 for src_file, dst_file in zip(api_src_files, api_dst_files):
-    shutil.copyfile(src_file, dst_file)
+    if (
+            not os.path.exists(dst_file)  # new file
+            or not filecmp.cmp(src_file, dst_file)  # file changed
+    ):
+        shutil.copyfile(src_file, dst_file)
